restructure plot package
