+++ autogalaxy/plot/plots/fit_imaging_plots.py
@@ -1,11 +1,11 @@
 import numpy as np
-from autoarray.plot.mat_wrap import plotters
+from autoarray.plot.plotter import plotter
 from autogalaxy.plot.plots import inversion_plots
-from autogalaxy.plot.mat_wrap import lensing_plotters
+from autogalaxy.plot.plotter import lensing_plotter
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_subplot_filename
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_subplot_filename
 def subplot_fit_imaging(fit, include=None, sub_plotter=None):
     number_subplots = 6
 
@@ -57,8 +57,8 @@ def subplots_of_all_galaxies(fit, include=None, sub_plotter=None):
             )
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_subplot_filename
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_subplot_filename
 def subplot_of_galaxy(fit, galaxy_index, include=None, sub_plotter=None):
     """Plot the model datas_ of an analysis, using the *Fitter* class object.
 
@@ -211,12 +211,12 @@ def individuals(
             )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def subtracted_image_of_galaxy(fit, galaxy_index, include=None, plotter=None):
     """Plot the model image of a specific plane of a lens fit.
 
-    Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -228,7 +228,7 @@ def subtracted_image_of_galaxy(fit, galaxy_index, include=None, plotter=None):
         The plane from which the model image is generated.
     """
 
-    if isinstance(plotter, lensing_plotters.Plotter):
+    if isinstance(plotter, lensing_plotter.Plotter):
         plotter = plotter.plotter_with_new_output(
             filename=plotter.output.filename + "_" + str(galaxy_index)
         )
@@ -262,12 +262,12 @@ def subtracted_image_of_galaxy(fit, galaxy_index, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def model_image_of_galaxy(fit, galaxy_index, include=None, plotter=None):
     """Plot the model image of a specific plane of a lens fit.
 
-    Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -277,7 +277,7 @@ def model_image_of_galaxy(fit, galaxy_index, include=None, plotter=None):
         The plane from which the model image is generated.
     """
 
-    if isinstance(plotter, lensing_plotters.Plotter):
+    if isinstance(plotter, lensing_plotter.Plotter):
         plotter = plotter.plotter_with_new_output(
             filename=plotter.output.filename + "_" + str(galaxy_index)
         )
@@ -291,12 +291,12 @@ def model_image_of_galaxy(fit, galaxy_index, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def image(fit, include=None, plotter=None):
     """Plot the image of a lens fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -316,12 +316,12 @@ def image(fit, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def noise_map(fit, include=None, plotter=None):
     """Plot the noise-map of a lens fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -340,12 +340,12 @@ def noise_map(fit, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def signal_to_noise_map(fit, include=None, plotter=None):
     """Plot the noise-map of a lens fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -364,12 +364,12 @@ def signal_to_noise_map(fit, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def model_image(fit, include=None, plotter=None):
     """Plot the model image of a fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -388,12 +388,12 @@ def model_image(fit, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def residual_map(fit, include=None, plotter=None):
     """Plot the residual-map of a lens fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -412,12 +412,12 @@ def residual_map(fit, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def normalized_residual_map(fit, include=None, plotter=None):
     """Plot the residual-map of a lens fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
@@ -436,12 +436,12 @@ def normalized_residual_map(fit, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def chi_squared_map(fit, include=None, plotter=None):
     """Plot the chi-squared-map of a lens fit.
 
-    Set *autogalaxy.datas.array.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.array.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
