+++ autogalaxy/plot/plots/galaxy_plots.py
@@ -1,15 +1,15 @@
-from autoarray.plot.mat_wrap import plotters
+from autoarray.plot.plotter import plotter
 from autoarray.structures import arrays
-from autogalaxy.plot.mat_wrap import lensing_plotters
+from autogalaxy.plot.plotter import lensing_plotter
 from autogalaxy.plot.plots import light_profile_plots, mass_profile_plots
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def image(galaxy, grid, positions=None, include=None, plotter=None):
     """Plot the image (e.g. the datas) of a galaxy, on a grid of (y,x) coordinates.
 
-    Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all innput parameters not described below.
+    Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all innput parameters not described below.
 
     Parameters
     -----------
@@ -29,12 +29,12 @@ def image(galaxy, grid, positions=None, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def convergence(galaxy, grid, positions=None, include=None, plotter=None):
     """Plot the convergence of a galaxy, on a grid of (y,x) coordinates.
 
-    Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all innput parameters not described below.
+    Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all innput parameters not described below.
 
     Parameters
     -----------
@@ -54,12 +54,12 @@ def convergence(galaxy, grid, positions=None, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def potential(galaxy, grid, positions=None, include=None, plotter=None):
     """Plot the potential of a galaxy, on a grid of (y,x) coordinates.
 
-     Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all innput parameters not described below.
+     Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all innput parameters not described below.
 
      Parameters
      -----------
@@ -79,12 +79,12 @@ def potential(galaxy, grid, positions=None, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def deflections_y(galaxy, grid, positions=None, include=None, plotter=None):
     """Plot the y component of the deflection angles of a galaxy, on a grid of (y,x) coordinates.
 
-    Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all innput parameters not described below.
+    Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all innput parameters not described below.
 
     Parameters
     -----------
@@ -109,12 +109,12 @@ def deflections_y(galaxy, grid, positions=None, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def deflections_x(galaxy, grid, positions=None, include=None, plotter=None):
     """Plot the x component of the deflection angles of a galaxy, on a grid of (y,x) coordinates.
 
-     Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all innput parameters not described below.
+     Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all innput parameters not described below.
 
      Parameters
      -----------
@@ -138,12 +138,12 @@ def deflections_x(galaxy, grid, positions=None, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def magnification(galaxy, grid, positions=None, include=None, plotter=None):
     """Plot the magnification of a galaxy, on a grid of (y,x) coordinates.
 
-     Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all innput parameters not described below.
+     Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all innput parameters not described below.
 
      Parameters
      -----------
@@ -164,8 +164,8 @@ def magnification(galaxy, grid, positions=None, include=None, plotter=None):
     )
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_labels
 def image_subplot(galaxy, grid, positions=None, include=None, sub_plotter=None):
 
     number_subplots = len(galaxy.light_profiles)
@@ -188,8 +188,8 @@ def image_subplot(galaxy, grid, positions=None, include=None, sub_plotter=None):
     sub_plotter.figure.close()
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_labels
 def convergence_subplot(galaxy, grid, positions=None, include=None, sub_plotter=None):
 
     number_subplots = len(galaxy.mass_profiles)
@@ -212,8 +212,8 @@ def convergence_subplot(galaxy, grid, positions=None, include=None, sub_plotter=
     sub_plotter.figure.close()
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_labels
 def potential_subplot(galaxy, grid, positions=None, include=None, sub_plotter=None):
 
     number_subplots = len(galaxy.mass_profiles)
@@ -236,8 +236,8 @@ def potential_subplot(galaxy, grid, positions=None, include=None, sub_plotter=No
     sub_plotter.figure.close()
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_labels
 def deflections_y_subplot(galaxy, grid, positions=None, include=None, sub_plotter=None):
 
     number_subplots = len(galaxy.mass_profiles)
@@ -260,8 +260,8 @@ def deflections_y_subplot(galaxy, grid, positions=None, include=None, sub_plotte
     sub_plotter.figure.close()
 
 
-@lensing_plotters.set_include_and_sub_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_sub_plotter
+@plotter.set_labels
 def deflections_x_subplot(galaxy, grid, positions=None, include=None, sub_plotter=None):
 
     number_subplots = len(galaxy.mass_profiles)
@@ -284,12 +284,12 @@ def deflections_x_subplot(galaxy, grid, positions=None, include=None, sub_plotte
     sub_plotter.figure.close()
 
 
-@lensing_plotters.set_include_and_plotter
-@plotters.set_labels
+@lensing_plotter.set_include_and_plotter
+@plotter.set_labels
 def contribution_map(galaxy, mask=None, positions=None, include=None, plotter=None):
     """Plot the summed contribution maps of a hyper_galaxies-fit.
 
-    Set *autogalaxy.datas.arrays.plotters.plotters* for a description of all input parameters not described below.
+    Set *autogalaxy.datas.arrays.plotter.plotter* for a description of all input parameters not described below.
 
     Parameters
     -----------
