+++ test_autogalaxy/unit/pipeline/test_visualizer.py
@@ -15,12 +15,7 @@ directory = path.dirname(path.abspath(__file__))
 
 @pytest.fixture(name="plot_path")
 def make_visualizer_plotter_setup():
-    return path.join(
-        "{}".format(path.dirname(path.realpath(__file__))),
-        "files",
-        "plot",
-        "visualizer",
-    )
+    return path.join("{}".format(directory), "files", "plot", "visualizer")
 
 
 @pytest.fixture()
