+++ test_autogalaxy/unit/plot/plotter/config/visualize/general.ini
@@ -1,0 +1,25 @@
+[general]
+backend=default
+visualize_every_update=10
+
+[units]
+in_kpc=True
+
+[include]
+origin=True
+mask=True
+grid=True
+border=True
+positions=True
+critical_curves=False
+caustics=False
+multiple_images=False
+light_profile_centres=True
+mass_profile_centres=True
+inversion_pixelization_grid=True
+inversion_grid=True
+inversion_border=True
+inversion_image_pixelization_grid=True
+parallel_overscan=True
+serial_prescan=True
+serial_overscan=True
