+++ test_autogalaxy/unit/plot/plotter/config/visualize/mat_obj/CausticsLiner.ini
@@ -1,0 +1,11 @@
+[figure]
+width=3
+style=--
+colors=w, g
+pointsize=21
+
+[subplot]
+width=7
+style=--
+colors=g
+pointsize=23
