+++ test_autogalaxy/unit/plot/plotter/config/visualize/mat_obj/CriticalCurvesLiner.ini
@@ -1,0 +1,11 @@
+[figure]
+width=2
+style=-
+colors=w, k
+pointsize=20
+
+[subplot]
+width=6
+style=-
+colors=b
+pointsize=22
