+++ test_autogalaxy/unit/plot/plotter/config/visualize/mat_obj/LightProfileCentreScatterer.ini
@@ -1,0 +1,9 @@
+[figure]
+size=10
+marker=+
+colors=k, r
+
+[subplot]
+size=15
+marker=.
+colors=b
