+++ test_autogalaxy/unit/plot/plotter/config/visualize/mat_obj/MassProfileCentreScatterer.ini
@@ -1,0 +1,9 @@
+[figure]
+size=11
+marker=x
+colors=r, k
+
+[subplot]
+size=16
+marker=o
+colors=k
