+++ test_autogalaxy/unit/plot/plotter/config/visualize/mat_obj/MultipleImagesScatterer.ini
@@ -1,0 +1,9 @@
+[figure]
+size=12
+marker=o
+colors=k, w
+
+[subplot]
+size=17
+marker=.
+colors=g
