+++ test_autogalaxy/unit/plot/plotter/config/visualize/plots.ini
@@ -1,0 +1,59 @@
+[dataset]
+subplot_dataset=True
+data=False
+noise_map=False
+psf=False
+uv_wavelengths=False
+inverse_noise_map=False
+signal_to_noise_map=False
+absolute_signal_to_noise_map=False
+potential_chi_squared_map=False
+
+[fit]
+all_at_end_png=True
+all_at_end_fits=True
+subplot_fit=True
+subplots_of_plane_fits=True
+data=False
+noise_map=False
+signal_to_noise_map=False
+model_data=False
+residual_map=False
+normalized_residual_map=False
+chi_squared_map=False
+subtracted_images_of_galaxies=False
+model_images_of_galaxies=False
+
+[inversion]
+subplot_inversion=True
+reconstructed_image=False
+reconstruction=False
+errors=False
+residual_map=False
+normalized_residual_map=False
+chi_squared_map=False
+regularization_weight_map=False
+interpolated_reconstruction=False
+interpolated_errors=False
+
+[hyper_galaxy]
+subplot_hyper_galaxy=True
+plot_hyper_model_image=True
+images=True
+
+[galaxy_fit]
+all_at_end_png=True
+all_at_end_fits=True
+subplot_galaxy_fit=True
+image=True
+noise_map=False
+model_image=False
+residual_map=False
+chi_squared_map=False
+
+[pdf]
+triangle=False
+1d_params=False
+
+[other]
+stochastic_histogram=False
