+++ test_autogalaxy/unit/plot/plotter/files/plotter/grids/radial_minimum.ini
